moved snackbar
* move snackbar into it's own component folder
